+++ examples/plugin_test.sh
@@ -85,12 +85,8 @@ echo NetScaler::System::Disk
 ./check_netscaler.pl --extra-opts=${section}@${1} -C above -o system -n disk0perusage,disk1perusage -w 75 -c 80
 echo
 
-echo NetScaler::HA::Status
-./check_netscaler.pl --extra-opts=${section}@${1} -C matches_not -o hanode -n hacurstatus -w YES -c YES
-echo
-
-echo NetScaler::HA::State
-./check_netscaler.pl --extra-opts=${section}@${1} -C matches_not -o hanode -n hacurstate -w UP -c UP
+echo NetScaler::HA:
+./check_netscaler.pl --extra-opts=${section}@${1} -C hastatus
 echo
 
 echo NetScaler::Perfdata::HTTP